Updated pass with DI registration
